+++ internal/sessions/repository/cache/cache.go
@@ -9,7 +9,7 @@ import (
 )
 
 var (
-	maxVersion uint8 = 255
+	maxVersion uint32 = 255
 )
 
 type SessionStorage struct {
@@ -22,19 +22,19 @@ func NewSessionStorage() *SessionStorage {
 	}
 }
 
-func (sessionStorage *SessionStorage) Add(login string, token string, version uint8) (err error) {
+func (sessionStorage *SessionStorage) Add(login string, token string, version uint32) (err error) {
 	sessionMap, hasUser := sessionStorage.cacheStorage.Get(login)
 	if !hasUser {
-		sesMap := make(map[string]uint8)
+		sesMap := make(map[string]uint32)
 		sesMap[token] = version
 		sessionStorage.cacheStorage.Set(login, sesMap, 0)
 		return nil
 	}
-	if _, hasSession := sessionMap.(map[string]uint8)[token]; hasSession {
+	if _, hasSession := sessionMap.(map[string]uint32)[token]; hasSession {
 		return myerrors.ErrItemsIsAlreadyInTheCache
 
 	}
-	sesMap := sessionMap.(map[string]uint8)
+	sesMap := sessionMap.(map[string]uint32)
 	sesMap[token] = version
 	sessionStorage.cacheStorage.Set(login, sesMap, 0)
 
@@ -43,8 +43,8 @@ func (sessionStorage *SessionStorage) Add(login string, token string, version ui
 
 func (sessionStorage *SessionStorage) DeleteSession(login string, token string) (err error) {
 	if sessionMapInterface, hasUser := sessionStorage.cacheStorage.Get(login); hasUser {
-		if _, hasSession := sessionMapInterface.(map[string]uint8)[token]; hasSession {
-			delete(sessionMapInterface.(map[string]uint8), token)
+		if _, hasSession := sessionMapInterface.(map[string]uint32)[token]; hasSession {
+			delete(sessionMapInterface.(map[string]uint32), token)
 			return nil
 		}
 		return myerrors.ErrNoSuchSessionInTheCache
@@ -54,11 +54,11 @@ func (sessionStorage *SessionStorage) DeleteSession(login string, token string) 
 
 func (sessionStorage *SessionStorage) Update(login string, token string) (err error) {
 	if sessionMapInterface, hasUser := sessionStorage.cacheStorage.Get(login); hasUser {
-		if _, hasSession := sessionMapInterface.(map[string]uint8)[token]; hasSession {
-			if (sessionMapInterface.(map[string]uint8))[token] == maxVersion {
+		if _, hasSession := sessionMapInterface.(map[string]uint32)[token]; hasSession {
+			if (sessionMapInterface.(map[string]uint32))[token] == maxVersion {
 				return myerrors.ErrTooHighVersion
 			}
-			(sessionMapInterface.(map[string]uint8))[token]++
+			(sessionMapInterface.(map[string]uint32))[token]++
 			return nil
 		}
 		return myerrors.ErrNoSuchSessionInTheCache
@@ -67,9 +67,9 @@ func (sessionStorage *SessionStorage) Update(login string, token string) (err er
 }
 
 func (sessionStorage *SessionStorage) CheckVersion(login string, token string,
-	usersVersion uint8) (hasSession bool, err error) {
+	usersVersion uint32) (hasSession bool, err error) {
 	if sessionMapInterface, hasUser := sessionStorage.cacheStorage.Get(login); hasUser {
-		if sessionMapInterface.(map[string]uint8)[token] == usersVersion {
+		if sessionMapInterface.(map[string]uint32)[token] == usersVersion {
 			return true, nil
 		}
 		return false, myerrors.ErrWrongSessionVersion
@@ -79,23 +79,23 @@ func (sessionStorage *SessionStorage) CheckVersion(login string, token string,
 
 func (sessionStorage *SessionStorage) HasSession(login string, token string) error {
 	if sessionMapInterface, hasUser := sessionStorage.cacheStorage.Get(login); hasUser {
-		if _, hasSession := sessionMapInterface.(map[string]uint8)[token]; hasSession {
+		if _, hasSession := sessionMapInterface.(map[string]uint32)[token]; hasSession {
 			return nil
 		}
 	}
 	return myerrors.ErrNoSuchUser
 }
 
-func (sessionStorage *SessionStorage) GetVersion(login string, token string) (version uint8, err error) {
+func (sessionStorage *SessionStorage) GetVersion(login string, token string) (version uint32, err error) {
 	if sessionMapInterface, hasUser := sessionStorage.cacheStorage.Get(login); hasUser {
-		return sessionMapInterface.(map[string]uint8)[token], nil
+		return sessionMapInterface.(map[string]uint32)[token], nil
 	}
 	return 0, myerrors.ErrNoSuchItemInTheCache
 }
 
 func (sessionStorage *SessionStorage) CheckAllUserSessionTokens(login string) error {
 	if sessionMapInterface, hasUser := sessionStorage.cacheStorage.Get(login); hasUser {
-		for token, version := range sessionMapInterface.(map[string]uint8) {
+		for token, version := range sessionMapInterface.(map[string]uint32) {
 			fmt.Println("token:", token, "version:", version)
 		}
 		return nil
